Optimizations to reduce callgraph creation time.
Lots of small optimizations here.  Three main categories:
1) Only doing lookups on maps/sets/lists one time and saving the result
instead of two or more times per operation.

2) Changing the graph datastructure to use a vector instead of set.
This change had a very significant impact on performance.  As best I
could tell from the algorithms there should be no worry about duplicate
entries from using a vector.

3) Significantly revised the graph merging algorithm for points-to.  In
particular, we no longer re-build the value_vertex_map on each merge,
rather we simply add the new entries to the existing map.  This also had
a notable impact on performance.

Some changes on the ICFG graph, but profiling did not indicate that this
was a sticking point.

The final profiling showed that the slowest, by far, part of the
callgraph generation at this point is the call to AA.alias in the
PointsToGraph constructor.  85% of the runtime was spent inside this
function.  I looked into being able to parallelize this section of code,
however the LLVM data structures used within
ProjectIRDB::preprocessModule do not allow for this.