PointsTo graph and and Callgraph optimizations.
Based on information from running via valgrind, targeted the slow points
which could easily be optimized.  The largest change is in the
PointsToGraph merge functions -- to avoid re-building the ValueVertexMap
from scratch for each merge.  Also, the llvm::User list for the
PointsToGraph VertexProperties was added to cache these results, as the
call to llvm::Value::users() can be time consuming and was invoked many
many times.

Some small improvements to LLVMBasedICFG::constructionWalker saved a bit
of time by avoiding multiple lookups on the VisitedFunctions and
FunctionVertexMap data structures.

Also, throughout the code the calls to boost::add_vertex and
boost::add_edge modified to pass in the new Vertex/Edge when adding,
instead of doing a lookup immediately after.

Finally, the backing data structures for the graphs were changed to
vectors. From my analysis this should be okay since the construction of
the graphs should already avoid duplicates.